requirements.txt
asgiref==3.2.10
Django==3.1.2
django-rest-multiple-models==2.1.3
djangorestframework==3.12.1
pytz==2020.1
sqlparse==0.4.1
